LayoutTests/imported/w3c/web-platform-tests/cors/allow-headers-expected.txt

Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=*
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=%20*%20%20
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=%09*
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=%20http%3A%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=%20http%3A%2F%2Flocalhost%3A8800%20%20%20%09%20
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=%09http%3A%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fwww1.localhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=%3A%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=ftp%3A%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%3A%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3Alocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=localhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Flocalhost%3A8800%3F
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Flocalhost%3A8800%2F
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Flocalhost%3A8800%20%2F
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Flocalhost%3A8800%23
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Flocalhost%3A8800%2523
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Flocalhost%3A8800%3A80
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Flocalhost%3A8800%2C%20*
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Flocalhost%3A8800%00
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=HTTP%3A%2F%2FLOCALHOST%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=HTTP%3A%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=-
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=**
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=%00*
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=*%00
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=%27*%27
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=%22*%22
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=*%20*
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=*http%3A%2F%2F*
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=*http%3A%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=*%20http%3A%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=*%2C%20http%3A%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=%00http%3A%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=null%20http%3A%2F%2Flocalhost%3A8800
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fexample.net
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=null
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Flocalhost%3A8800%2Fcors%2Fallow-headers.htm
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Flocalhost%3A8800%2Fcors%2F
Blocked access to external URL http://www1.localhost:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fwww1.localhost%3A8800%2Fcors%2F
Access-Control-Allow-Headers handling


FAIL Allow origin: * The operation was aborted.
FAIL Allow origin: _*__ The operation was aborted.
FAIL Allow origin: [tab]* The operation was aborted.
FAIL Allow origin: http://localhost:8800 The operation was aborted.
FAIL Allow origin: _http://localhost:8800 The operation was aborted.
FAIL Allow origin: _http://localhost:8800___[tab]_ The operation was aborted.
FAIL Allow origin: [tab]http://localhost:8800 The operation was aborted.
FAIL Disallow origin: http://www1.localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: //localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: ://localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: ftp://localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http:://localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http:/localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http:localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http://localhost:8800? assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http://localhost:8800/ assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http://localhost:8800 / assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http://localhost:8800# assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http://localhost:8800%23 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http://localhost:8800:80 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http://localhost:8800, * assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http://localhost:8800\0 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: HTTP://LOCALHOST:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: HTTP://localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: - assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: ** assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: \0* assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: *\0 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: '*' assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: "*" assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: * * assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: *http://* assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: *http://localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: * http://localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: *, http://localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: \0http://localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: null http://localhost:8800 assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http://example.net assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: null assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin:  assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http://localhost:8800/cors/allow-headers.htm assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http://localhost:8800/cors/ assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
FAIL Disallow origin: http://www1.localhost:8800/cors/ assert_throws: send function "function () { client.send() }" threw object "AbortError (DOM Exception 20): The operation was aborted." that is not a DOMException NetworkError: property "code" is equal to 20, expected 19
